+++ Changelog.md
@@ -1,9 +1,12 @@
 # Changelog
 
 ## [Unreleased]
+## [3.15.0] - 2026-07-15
+- Added credit expiry mail processing
+
 ## [3.14.0] - 2026-07-14
 ### Changed
-- Added a way to trac credit issuer.
+- Added a way to track credit issuer.
 
 ## [3.13.1] - 2026-07-06
 ### Changed